fix(gatsby): nodeModel.findAll supports v5 sort argument structure
